+++ deployment/deploy.sh
@@ -1,87 +1,146 @@
 #/bin/bash
 
-# Expects a commit pushed with these items:
+# To trigger a production deploy push a 
+# commit with these items: 
 #
 # version: x.y.z
 # Changelog: 
 
 set -e
 
-# Updating Changelog from commit message
-GIT_COMMIT_MSG=$(git log --format=%B -n 1 $CIRCLE_SHA1)
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+NC='\033[0m'
 
-VERSION=$(echo "$GIT_COMMIT_MSG" | grep version | cut -f 2 -d " ")
-
-DATE=$(date "+%Y-%m-%d")
-
-if [[ "$GIT_COMMIT_MSG" != *"version"* ]]; then
-    echo "Version not found in commit message - Not deploying"
-    exit 0
-fi
-
-echo "Extracted version $VERSION "
-
-# ToDo: Enable this section after adding public deploy key to build project
-#CHANGELOG=$(cat $GIT_COMMIT_MSG | awk '/Changelog/{y=1;next}y')
-#
-#INSERT="\n## [v$VERSION] - $date\n$CHANGELOG"
-#
-#if [ -z "$CHANGELOG" ]; then
-#    echo "Changelog not found in commit message - exiting"
-#    exit 1
-#fi
-#
-#cat <<EOF >add.txt
-#
-#$INSERT
-#$CHANGELOG
-#EOF
-
-## Update CHANGELOG.md
-#sed -i '.bak' '/\#\# \[VERSION\] - unreleased/r add.txt' CHANGELOG.md
-
-#sed -i -e "s/## \[VERSION\] - unreleased/## [$VERSION] - $DATE/" CHANGELOG.md
-#perl -i -pe '$_ = "\n## [VERSION] - unreleased\n\n" if $. ==4' CHANGELOG.md
-
-echo "Bumping versions ..."
-sed -i -e "s/version = '.*';$/version = '$VERSION';/" src/0_config.js
-sed -i -e "s/version = '.*';$/version = '$VERSION';/" test/web-config.js
-
-sed -i -e "s/\"version\":.*$/\"version\": \"$VERSION\",/" package.json
-sed -i -e "s/\"build\":.*$/\"build\": \"$VERSION\"/" package.json
-
-sed -i -e "s/\"version\":.*$/\"version\": \"$VERSION\",/" bower.json
-sed -i -e "s/\"build\":.*$/\"build\": \"$VERSION\"/" bower.json
-
-echo "make release ..."
-make release
-
-
-# ToDo: Enable this section after adding public deploy key to build project
-#chmod 600 ~/deploy/Key.pem
-#git config --global user.email "buildbot@branch.io" && git config --global user.name "Build Bot"
-#git config --global push.default simple
-
-#echo "Commiting changes back to repo"
-#git add dist/build.min.js.gz
-#git commit -am "Tagging release $VERSION [ci skip]"
-#git add $HOME/$CIRCLE_PROJECT_REPONAME/CHANGELOG.md
-#git tag $VERSION
+#--------------------------------------------------------------------------------------------
+# Production Deploy
+#--------------------------------------------------------------------------------------------
 
 if [ "$CIRCLE_BRANCH" == 'production' ]; then
-  echo "Pushing to S3: branch-cdn ..."
+  
+  echo -en "${GREEN}Production Release...${NC}\n"
+
+  GIT_COMMIT_MSG=$(git log --format=%B -n 1 $CIRCLE_SHA1)
+  
+  VER=$(echo "$GIT_COMMIT_MSG" | grep version | cut -f 2 -d " ")
+ 
+  VERSION=$(echo $VER|tr -d '\r')
+ 
+  DATE=$(date "+%Y-%m-%d")
+  
+  if [[ "$GIT_COMMIT_MSG" != *"version"* ]]; then
+      echo "Version not found in commit message - Not deploying"
+      exit 0
+  fi
+  
+  echo -en "${GREEN} Extracted version $VERSION ${NC}\n"
+ 
+  # Expect a Changelog in commit message
+  CHANGELOG=$(echo "$GIT_COMMIT_MSG" | awk '/Changelog/{y=1;next}y')
+  INSERT="## [v$VERSION] - $DATE"
+
+  if [ -z "$CHANGELOG" ]; then
+      echo "Changelog not found in commit message - Not deploying"
+      exit 0
+  fi
+
+  echo -en "\n${GREEN}Extracted Changelog:\n$INSERT\n$CHANGELOG\n${NC}\n"
+
+cat <<EOF >add.txt
+$CHANGELOG
+EOF
+
+  # Update CHANGELOG.md
+  sed -i '/\#\# \[VERSION\] - unreleased/r add.txt' CHANGELOG.md
+
+  sed -i -e "s/## \[VERSION\] - unreleased/## [$VERSION] - $DATE/" CHANGELOG.md
+  perl -i -pe '$_ = "\n## [VERSION] - unreleased\n\n" if $. ==4' CHANGELOG.md
+ 
+  echo -en "${GREEN}Bumping versions ...${NC}\n"
+  sed -i -e "s/version = '.*';$/version = '$VERSION';/" src/0_config.js
+  sed -i -e "s/version = '.*';$/version = '$VERSION';/" test/web-config.js
+  
+  sed -i -e "s/\"version\":.*$/\"version\": \"$VERSION\",/" package.json
+  sed -i -e "s/\"build\":.*$/\"build\": \"$VERSION\"/" package.json
+  
+  sed -i -e "s/\"version\":.*$/\"version\": \"$VERSION\",/" bower.json
+  sed -i -e "s/\"build\":.*$/\"build\": \"$VERSION\"/" bower.json
+
+  echo -en "${GREEN}make release...${NC}\n"
+  make release
+
+  rm add.txt
+
+  echo -en "${GREEN}Commiting changes back to repo${NC}\n"
+  git config --global user.email "buildbot@branch.io" && git config --global user.name "Build Bot"
+  git config --global push.default simple
+  git commit -am "Pushing release $VERSION [ci skip]"
+
+  echo -en "${GREEN}Pushing to S3: branch-cdn ...${NC}\n"
   aws s3 cp --content-type="text/javascript" --content-encoding="gzip" dist/build.min.js.gz s3://branch-cdn/branch-$VERSION.min.js --acl public-read
   aws s3 cp --content-type="text/javascript" --content-encoding="gzip" dist/build.min.js.gz s3://branch-cdn/branch-latest.min.js --acl public-read
   aws s3 cp --content-type="text/javascript" --content-encoding="gzip" dist/build.min.js.gz s3://branch-cdn/branch-v2.0.0.min.js --acl public-read
   aws s3 cp example.html s3://branch-cdn/example.html --acl public-read
   
   # Invalidate cache at CDN
-  echo "Invalidating cloudfrond distribution for WebSDK ..."
+  echo -en "${GREEN}Invalidating cloudfrond distribution for WebSDK ...${NC}\n"
   aws configure set preview.cloudfront true
-  aws cloudfront create-invalidation --distribution-id E10P37NG0GMER --paths /
+  aws cloudfront create-invalidation --distribution-id E10P37NG0GMER --paths /branch-latest.min.js /example.html /branch-v2.0.0.min.js
+
+  echo -en "${GREEN}Pushing git tags${NC}\n"
+  git tag v$VERSION
+  git push origin v$VERSION
+
+  echo -en "${GREEN}npm publish ...${NC}\n"
+  npm publish
+
+  # Reset
+  echo -en "${GREEN}make clean ...${NC}\n"
+  make clean
+  make
+  git commit -am "Resetting to HEAD [ci skip]"
+
+  echo -en "${GREEN}Updating production files ...${NC}\n"
+  git push origin $CIRCLE_BRANCH
+
+  # Push back to master
+  echo -en "${GREEN}Updating Master files ...${NC}\n"
+  rm -rf /tmp/$CIRCLE_PROJECT_REPONAME
+  git clone git@github.com:BranchMetrics/$CIRCLE_PROJECT_REPONAME.git /tmp/$CIRCLE_PROJECT_REPONAME
+
+  pushd /tmp/$CIRCLE_PROJECT_REPONAME 
+
+  git merge origin/production -m "Merge Production [ci skip]"
+  git push origin master
+  popd
+
+  # Send an update to slack channels
+
+  DEPLOY_IMG=http://workshops.lewagon.org/assets/landing-2/deploy-button-5068ec2c575492ba428569111afe3ce6.jpg
+  echo -en "${GREEN}Sending update to slack ...${NC}\n"
+  #uncomment to send updates to int-eng
+  #slackcli -t $SLACK_TOKEN -h int-eng -m $MESSAGE -u websdk-deploy -i $DEPLOY_IMG
+  slackcli -t $SLACK_TOKEN -h web-sdk -m "$CIRCLE_USERNAME deployed WebSDK v$VERSION" -u websdk-deploy -i $DEPLOY_IMG
+  
+  echo "Please update the javascript version in https://github.com/BranchMetrics/documentation/edit/master/ingredients/web_sdk/_initialization.md"
+
+  echo "Check the following links are valid:"
+  echo "- https://cdn.branch.io/branch-$VERSION.min.js"
+  echo "- https://cdn.branch.io/example.html using the right version number $VERSION"
+  echo "- https://www.npmjs.com/package/branch-sdk using the right version number $VERSION"
+
+#--------------------------------------------------------------------------------------------
+# Master Deploy
+#--------------------------------------------------------------------------------------------
 
 elif [ "$CIRCLE_BRANCH" == 'master' ]; then
-  echo "Pushing to S3: branch-builds ..."
+
+  echo -en "${GREEN}QA Release...${NC}\n"
+
+  echo -en "${GREEN}make release ...${NC}\n"
+  make release
+
+  echo -en "${GREEN}Pushing to S3: branch-builds ...${NC}\n"
   aws s3 cp --content-type="text/javascript" --content-encoding="gzip" dist/build.min.js.gz s3://branch-builds/websdk/branch-$VERSION.min.js --acl public-read
   aws s3 cp --content-type="text/javascript" --content-encoding="gzip" dist/build.min.js.gz s3://branch-builds/websdk/branch-latest.min.js --acl public-read
   aws s3 cp --content-type="text/javascript" --content-encoding="gzip" dist/build.min.js.gz s3://branch-builds/websdk/branch-v2.0.0.min.js --acl public-read
@@ -89,39 +148,16 @@ elif [ "$CIRCLE_BRANCH" == 'master' ]; then
   aws s3 cp example.html s3://branch-builds/websdk/example.html --acl public-read
 
 else
-    echo "No associated bucket to $CIRCLE_BRANCH - not Deploying"
+    echo -en "${GREEN}No associated bucket to $CIRCLE_BRANCH - not Deploying${NC}\n"
     exit 0
-fi	
+fi
 
-#echo "npm publish ..."
-#npm publish
-
-# Reset
-echo "make clean ..."
-make clean
-make
-
-# Cleaning up backup files
-echo "Cleaning up backup files ..."
-rm -f bower.json-e CHANGELOG.md-e package.json-e src/0_config.js-e test/web-config.js-e
-
-#echo "Pushing changes back to repo ..."
-#git push 
-#git push origin $VERSION
-
-# Send an update to slack channels
-DEPLOY_IMG=http://workshops.lewagon.org/assets/landing-2/deploy-button-5068ec2c575492ba428569111afe3ce6.jpg
-echo "Sending update to slack ..."
-#slackcli -t $SLACK_TOKEN -h int-eng -m "$CIRCLE_USERNAME Deploying WedSDK:$CIRCLE_BRANCH v$VERSION" -u websdk-deploy -i $DEPLOY_IMG
-slackcli -t $SLACK_TOKEN -h web-sdk -m "$CIRCLE_USERNAME Deployed WedSDK:$CIRCLE_BRANCH v$VERSION" -u websdk-deploy -i $DEPLOY_IMG
+# Rollbar updates
+if [ "$CIRCLE_BRANCH" == 'production' ] || [ "$CIRCLE_BRANCH" == 'master' ] ; then
+    pip install requests
+    deployment/rollbar.py
+fi
 
 # Exit prompts
-
-echo "Done script ..."
-echo "Please update the javascript version in https://github.com/BranchMetrics/documentation/edit/master/ingredients/web_sdk/_initialization.md"
-
-echo "Check the following links are valid:"
-echo "- https://cdn.branch.io/branch-$VERSION.min.js"
-echo "- https://cdn.branch.io/example.html using the right version number $VERSION"
-echo "- https://www.npmjs.com/package/branch-sdk using the right version number $VERSION"
+echo -en "${GREEN}Done script ...${NC}\n"